# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeMail 菜单化后端 API",
    "version": "2.0.0",
    "description": "WeMail 后端 API 已按管理后台左侧菜单分组。旧 /auth、/admin、/api/mailboxes、/api/messages、/api/outbound、/api/keys、/api/telegram 路径不再保留。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前 Worker 服务根路径"
    }
  ],
...
  "paths": {
...
    "/api/telegram/subscription": {
      "get": {
        "tags": [
          "Telegram"
        ],
        "summary": "获取 Telegram 订阅",
        "operationId": "getTelegramSubscription",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "订阅状态"
          }
        }
      },
      "put": {
        "tags": [
          "Telegram"
        ],
        "summary": "更新 Telegram 订阅",
        "operationId": "updateTelegramSubscription",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "订阅状态"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/system/features": {
      "get": {
        "tags": [
          "系统设置"
        ],
        "summary": "获取功能开关",
        "operationId": "getSystemFeatures",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "功能开关"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "tags": [
          "系统设置"
        ],
        "summary": "更新功能开关",
        "operationId": "updateSystemFeatures",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "功能开关"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wemail_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "请求失败",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}